kartoteka_server/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
httpx==0.27.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
SQLAlchemy>=1.4.51
passlib[bcrypt]==1.7.4
bcrypt>=3.2,<4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
jinja2==3.1.4
beautifulsoup4==4.12.3
apscheduler==3.10.4
pytest
pytest-mock
Pillow==10.1.0
imagehash==4.3.1
openai>=1.0.0